requirements.txt
pydantic
aiohttp
yarl
lxml
aiogram
python-dotenv
maigret
holehe
selenium
webdriver-manager
beautifulsoup4
requests